+++ odoo/local-src/connector_esb/tests/__init__.py
@@ -14,6 +14,7 @@ from . import test_export_special_promotion
 from . import test_export_product_price
 from . import test_export_buyx_gety
 from . import test_export_stock_update
+from . import test_export_stock_update_single
 from . import test_export_document_zip
 from . import test_product_supplierinfo_esbflux
 from . import test_product_supplierinfo_esbflux_purge